start tests, start other bindings
Started writing tests for the project. I'm starting with integration
tests to verify the semantics of regexes compiled by the compiler.
Later, I should add unit tests to verify the behavior of some of the
components, but that seems difficult because I'm dealing with
complicated graphs and compiled stuff.

I also started work on other language bindings. The first one is the
C/C++ dynamic loading binding. This uses libdl to dynamically load
functions from a shared library compiled by aotrc. This can be used
to dynamically run a regex without calling the function manually.