parser: rewrite parser with Bison/Flex
Antlr was good, but its issue is that it requires a
lot of dependencies (runtime, cmake stuff, etc). I
decided it would likely be easier to work with a
Bison/Flex parser generator system because all I need
is a flex and bison executable. Plus, there's much
better Flex/Bison cmake support (imo).

While it will take time for the parser to become as
fully featured as the antlr parser, I think this move
will be good in the long run.