compiler: add program mode structure for initializing program data
Added a new thing called program mode. THe purpose of program mode is to
be used in the program constructor to initialize data based on the kind
of program being compiled. The program mode constructs the function name,
function type, and program state for any program of a given type.

Arguably, this functionality should be integrated into the program class.
Putting lots of logic into a constructor is kind of a challenge. But, this
is designed to try to prevent code copying.