compiler: add search program, search program state, and search insts
Created a special type of program called the search program. The search program
searches an input string and finds sub-matches. It also uses a special version
of the program state that holds an additional variable. It also has a couple of
special instructions that only work with SearchProgramState.

In the future, sub-matching should show where the match actually occurs. For now,
it just says yes or no.