compiler: add submatching (#2)
This commit is responsible for implementing sub matching. This type of program searches the subject string to see if there is a substring that matches the given pattern. This sub match only tells if there *is* a submatch, not where it is. Searching will do that, which will come in a different commit.

Added a new thing called program mode. The purpose of program mode is to
be used in the program constructor to initialize data based on the kind
of program being compiled. The program mode constructs the function name,
function type, and program state for any program of a given type.

Arguably, this functionality should be integrated into the program class.
Putting lots of logic into a constructor is kind of a challenge. But, this
is designed to try to prevent code copying.

Also, created a special type of program called the search program. The search program
searches an input string and finds sub-matches. It also uses a special version
of the program state that holds an additional variable. It also has a couple of
special instructions that only work with SearchProgramState.

In the future, sub-matching should show where the match actually occurs. For now,
it just says yes or no.

Also added the pass system is used for running peephole optimization/manipulation on
a program. A pass is run over the instructions in an instruction vector
and can manipulate the instructions therein.