compiler: rewrite compiler internals to be more modular (#6)
Redesign compiler internals to be more modular. Now, compilation is done is phases:
Phase 1: DFA -> HIR
A DFA is converted into a sequence of instructions via a DFACompiler. A DFA compiler builds out a preable and
states.

Phase 2: HIR -> LLVM IR
Converts a HIR program to LLVM IR using an InstructionTranslator to translate each HIR instruction to LLVM IR.

The purpose of this design change is to try to make the compiler more modular. The old design was confusing as it relied too much on "passes," which really made it difficult to follow what's happening. There is still a ways to go on improving the compiler, but it's better now.